fix issues with rgb_t::white in a global declaration.
- const colors appear to be uninitialized when used in a global
declaration, and appear as black. this appears to be a problem
with static initialization on some compilers, where the const
colors have not been initialized before used in a static
declaration. change const initializers to #defines.

most obviously, this fixes missing rabbits in vicdual:carnival
and issues with other vicdual drivers using the color palette.